add Libertine fonts (fork of Linux Libertine/Biolinum)
The 'Linux Libertine' family is a great font family but the original
developer has been inactive for several years and the bug list is piling
up. A fork project know as just "Libertine' has many of the
long-standing issues fixed and has expanded the family with a math font.
The font faces have been renamed to not conflict with the original
project.