Adds the name stanza for all fonts
Currently, some fonts won't install, as reported in #437

I made a little shell script based on the script used to report the
problematic fonts to fix them.

It is probably not the best solution, but at least users will be able to
install the fonts.

My script is:

```sh
for font in $(grep --files-without-match --extended-regexp "\s+name'.*'" *); do
  name="${${${font#font-}%.rb}/-/ }"
  echo "$font -> $name"
  sed -i '' -e $'1 s/do/do\\\n  name '"'$name'"'/' "$font"
done
```