Flesh out documentation for token generation
Note that `otfinfo --family` works for most cases, but some fonts encode
a computer readable name here yet the actual human readable name of the
font is elsewhere. It can be found in the "Preferred name" section of
`otfinfo --info` but that is hard to parse. Using `fc-query -f '%{family}'`
is more consistent with what is actually reported to system applications
about installed fonts.

The extra step in removing meaningless strings is to account for the
handful of fonts that insist on cluttering the family name field with
redundant information such as "{name} Font" "{name} TrueType Font
Family".