Added `Doulos SIL` font v5.000
Doulos SIL is a font that supports  wide range of languages that use
the Latin and Cyrillic scripts, whether used for phonetic or
orthographic needs. Linguists appreciate the wide range of characters
and symbols useful in their work. This font makes use of
state-of-the-art font technologies to support complex typographic
issues, such as the need to position arbitrary combinations of base
glyphs and diacritics optimally.