skip checking sha256 digest for jf-open-huninn
The official download link from https://justfont.com/huninn/ refers
directly to the zip file generated out of the current "master" branch,
instead of a tagged released. This means that the sha256 digest in the
cask can be easily invalided by, say, a typo fix.

Given this, perhaps it's better to skip checking the sha256 digest.